requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
langchain
langchain-community
langchain-huggingface
sentence-transformers
chromadb
rank-bm25
transformers
torch
authlib
httpx
itsdangerous
cryptography
langfuse
networkx
fastembed

# Phase A — New dependencies
pinecone>=3.0
pinecone-text>=0.9
arq>=0.26
redis[asyncio]>=5.0
slowapi>=0.1.9
neo4j>=5.0
aiohttp>=3.9
